requirements.txt
pandas==2.1.4
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0 